Adjust tasks num when removing one, fixes #530

+++ cmscontrib/RemoveTask.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 # Contest Management System - http://cms-dev.github.io/
-# Copyright © 2013 Stefano Maggiolo <s.maggiolo@gmail.com>
+# Copyright © 2013-2016 Stefano Maggiolo <s.maggiolo@gmail.com>
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -26,9 +26,18 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import argparse
+import sys
 
 from cms import utf8_decoder
 from cms.db import SessionGen, Task
+
+
+def ask(task_name):
+    print("This will delete task `%s' and all related data, "
+          "including submissions. Are you sure? [y/N] "
+          % task_name, end='')
+    ans = sys.stdin.readline().strip().lower()
+    return ans in ["y", "yes"]
 
 
 def remove_task(task_name):
@@ -36,9 +45,25 @@ def remove_task(task_name):
         task = session.query(Task)\
             .filter(Task.name == task_name).first()
         if not task:
+            print("No task called `%s' found." % task_name)
             return
+        if not ask(task_name):
+            print("Not removing task `%s'." % task_name)
+            return
+        num = task.num
+        contest_id = task.contest_id
         session.delete(task)
         session.commit()
+        # Keeping the tasks' nums to the range 0... n - 1.
+        if contest_id is not None:
+            following_tasks = session.query(Task)\
+                .filter(Task.contest_id == contest_id)\
+                .filter(Task.num > num)\
+                .all()
+            for task in following_tasks:
+                task.num -= 1
+            session.commit()
+        print("Task `%s' removed." % task_name)
 
 
 def main():
